requirements.txt
euclid3==0.1
ply==3.11
prettytable==0.7.2
pypng==0.0.19
solidpython==1.1.3
